removing search from the hot board games page
