+++ board_games_companion/lib/pages/hot_board_games/search_results.dart
